Rearrange files - RSC-1077
