+++ lib/src/components/NxDescriptionList/types.ts
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2019-present Sonatype, Inc.
+ * This program and the accompanying materials are made available under
+ * the terms of the Eclipse Public License 2.0 which accompanies this
+ * distribution and is available at https://www.eclipse.org/legal/epl-2.0/.
+ */
+import { ButtonHTMLAttributes, HTMLAttributes, ReactElement } from 'react';
+import { ValidationMap } from 'prop-types';
+
+export interface BaseList<T extends HTMLElement = HTMLLIElement> extends HTMLAttributes<T> {
+  selected?: boolean | null,
+  disabled?: boolean | null,
+  buttonClassName?: string | null,
+  buttonAttributes?: ButtonHTMLAttributes<HTMLButtonElement> | null
+}
+
+export interface NxDescriptionListButtonItemProps extends BaseList<HTMLDivElement> {
+  children: ReactElement[]
+}
+
+export const nxDescriptionListButtonItemPropTypes: ValidationMap<NxDescriptionListButtonItemProps> = {
+};
