+++ lib/src/components/NxList/types.ts
@@ -9,18 +9,10 @@ import {
   LiHTMLAttributes,
   ValidationMap,
   ReactNode,
-  ReactElement,
   ButtonHTMLAttributes,
   AnchorHTMLAttributes
 } from 'react';
 import PropTypes from 'prop-types';
-
-export interface BaseList<T extends HTMLElement = HTMLLIElement> extends HTMLAttributes<T> {
-  selected?: boolean | null,
-  disabled?: boolean | null,
-  buttonClassName?: string | null,
-  buttonAttributes?: ButtonHTMLAttributes<HTMLButtonElement> | null
-}
 
 export interface NxListProps extends HTMLAttributes<HTMLUListElement> {
   bulleted?: boolean | null;
@@ -66,11 +58,4 @@ export const nxListLinkItemPropTypes: ValidationMap<NxListLinkItemProps> = {
   href: PropTypes.string.isRequired,
   anchorClassName: PropTypes.string,
   anchorAttributes: PropTypes.object
-};
-
-export interface NxDescriptionListButtonItemProps extends BaseList<HTMLDivElement> {
-  children: ReactElement[]
-}
-
-export const nxDescriptionListButtonItemPropTypes: ValidationMap<NxDescriptionListButtonItemProps> = {
 };
